+++ src/components/guestbooks/GuestBookContainer.jsx
@@ -3,27 +3,21 @@ import GuestBookHeader from './GuestBookHeader'
 import GuestBookBody from './GuestBookBody'
 import { dummyGuestBooks } from '../../constants/dummyGuestBooks'
 import { GuestBookContextForData, GuestBookContextForMethod } from '../contexts/GuestBookContexts'
+import { getCurrentDatetime } from '../../utils/getCurrentDatetime'
 
-export   const getCurrentDateTime = () => {
-  const now = new Date();
-  const year = now.getFullYear();
-  const month = String(now.getMonth() + 1).padStart(2, '0');
-  const day = String(now.getDate()).padStart(2, '0');
-  const hours = String(now.getHours()).padStart(2, '0');
-  const minutes = String(now.getMinutes()).padStart(2, '0');
-  const seconds = String(now.getSeconds()).padStart(2, '0');
-  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-} 
+
 const GuestBookContainer = () => {
+ 
     const [guestbooks, setGuestbooks] = useState(dummyGuestBooks);
+
 
     const addGuestBookHandler = ({nickname,content}) => {
         console.log("여기서닉네임",nickname)
         const newGuestBook = {
-            id : self.crypto.randomUUID(),
+            id : guestbooks.length + 1 ,
             nickname,
             content,
-            dateTime: getCurrentDateTime()
+            datetime: getCurrentDatetime()
         }
         setGuestbooks ([...guestbooks, newGuestBook])
     }
